requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
requests==2.28.2
steam==1.4.4
bufflogin>=0.0.4
pysteamauth>=1.0.0